requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorboard>=2.13.0

# Data Processing
numpy>=1.24.0
Pillow>=10.0.0
opencv-python>=4.8.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Evaluation & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0

